requirements.txt
gradio==4.27.0
langchain==0.1.16
langchain_community==0.0.34
langchain_core==0.1.45
langchain_text_splitters==0.0.1
ollama==0.1.8
webvtt_py==0.4.6
tiktoken==0.6.0